Native vs. function cleanup
- Use env variable to toggle unique check for postgres and spec
- Move native functionality to the include so it's available to jruby as well
